Restructured code, to separate different primitives into separate files Added useUserAgent primitive
